# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface QuizQuestion {
  question: string;
  /**
   * @minItems 2
   * @maxItems 10
   */
  options: string[];
  correctOptionIndex: number;
  explanation?: string;
}

export interface Quiz {
  id: number;
  title: string;
  sourceContent: string;
  questions: QuizQuestion[];
  questionCount: number;
  createdAt: string;
  updatedAt: string;
  postedToTelegram: boolean;
  telegramChannel?: string | null;
}

export interface GenerateQuizBody {
  title?: string;
  content: string;
  imageBase64?: string | null;
  /**
   * @minimum 1
   * @maximum 50
   */
  questionCount?: number;
  language?: string;
}

export interface UpdateQuizBody {
  title?: string;
  questions?: QuizQuestion[];
}

export interface PostToTelegramBody {
  botToken: string;
  channelId: string;
  /** If null, posts all questions */
  questionIndex?: number | null;
}

export interface TelegramPostResult {
  success: boolean;
  postedCount: number;
  messageIds: number[];
}

export interface ExportResult {
  data: string;
  filename: string;
  format: string;
}

export interface QuizStats {
  totalQuizzes: number;
  totalQuestions: number;
  postedToTelegram: number;
  recentQuizzes: Quiz[];
}

export interface ValidateBotBody {
  botToken: string;
}

export interface BotInfo {
  valid: boolean;
  username?: string | null;
  firstName?: string | null;
}

export type ExportQuizParams = {
  format: ExportQuizFormat;
};

export type ExportQuizFormat =
  (typeof ExportQuizFormat)[keyof typeof ExportQuizFormat];

export const ExportQuizFormat = {
  csv: "csv",
  json: "json",
} as const;