# .migration-backup/lib/api-zod/src/generated/types/generateQuizBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateQuizBody {
  title?: string;
  content?: string;
  imageBase64?: string | null;
  /**
   * @minimum 1
   * @maximum 50
   */
  questionCount?: number;
  language?: string;
}