# .migration-backup/lib/api-zod/src/generated/types/postToTelegramBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PostToTelegramBody {
  botToken: string;
  channelId: string;
  /** If null, posts all questions */
  questionIndex?: number | null;
}