# .migration-backup/lib/api-zod/src/generated/types/quiz.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from "./quizQuestion";

export interface Quiz {
  id: number;
  title: string;
  sourceContent: string;
  questions: QuizQuestion[];
  questionCount: number;
  createdAt: Date;
  updatedAt: Date;
  postedToTelegram: boolean;
  telegramChannel?: string | null;
}