# .migration-backup/lib/api-zod/src/generated/types/quizQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizQuestion {
  question: string;
  /**
   * @minItems 2
   * @maxItems 10
   */
  options: string[];
  correctOptionIndex: number;
  explanation?: string;
}