# .migration-backup/lib/api-zod/src/generated/types/updateQuizBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizQuestion } from "./quizQuestion";

export interface UpdateQuizBody {
  title?: string;
  questions?: QuizQuestion[];
}